+++ Configuration.sln
@@ -41,9 +41,9 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configu
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configuration.Binder.Test", "test\Microsoft.Framework.Configuration.Binder.Test\Microsoft.Framework.Configuration.Binder.Test.xproj", "{AE6FFE9B-6378-4D57-AA24-7D257F18B235}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configuration.File", "src\Microsoft.Framework.Configuration.File\Microsoft.Framework.Configuration.File.xproj", "{881E7CBC-492C-47C5-98A6-61DD1C753EE6}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configuration.FileExtensions", "src\Microsoft.Framework.Configuration.FileExtensions\Microsoft.Framework.Configuration.FileExtensions.xproj", "{881E7CBC-492C-47C5-98A6-61DD1C753EE6}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configuration.File.Test", "test\Microsoft.Framework.Configuration.File.Test\Microsoft.Framework.Configuration.File.Test.xproj", "{F7932F19-EB68-4C52-9CD1-3B51E48C2337}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Configuration.FileExtensions.Test", "test\Microsoft.Framework.Configuration.FileExtensions.Test\Microsoft.Framework.Configuration.FileExtensions.Test.xproj", "{F7932F19-EB68-4C52-9CD1-3B51E48C2337}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
