+++ src/Microsoft.Framework.Configuration.FileExtensions/Microsoft.Framework.Configuration.FileExtensions.xproj
