+++ src/Microsoft.Framework.Configuration.FileExtensions/project.json
