+++ src/Microsoft.Framework.Configuration.Ini/IniConfigurationExtensions.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using Microsoft.Framework.Configuration.Helper;
 using Microsoft.Framework.Configuration.Ini;
 
 namespace Microsoft.Framework.Configuration
@@ -53,7 +52,7 @@ namespace Microsoft.Framework.Configuration
                 throw new ArgumentException(Resources.Error_InvalidFilePath, nameof(path));
             }
 
-            var fullPath = ConfigurationHelper.ResolveConfigurationFilePath(configuration, path);
+            var fullPath = configuration.GetConfigurationFilePath(path);
 
             if (!optional && !File.Exists(fullPath))
             {
