+++ src/Microsoft.Framework.Configuration.Json/JsonConfigurationExtensions.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using Microsoft.Framework.Configuration.Helper;
 using Microsoft.Framework.Configuration.Json;
 
 namespace Microsoft.Framework.Configuration
@@ -58,7 +57,7 @@ namespace Microsoft.Framework.Configuration
                 throw new ArgumentException(Resources.Error_InvalidFilePath, nameof(path));
             }
 
-            var fullPath = ConfigurationHelper.ResolveConfigurationFilePath(configuration, path);
+            var fullPath = configuration.GetConfigurationFilePath(path);
 
             if (!optional && !File.Exists(fullPath))
             {
