+++ src/Microsoft.Framework.Configuration.Json/project.json
@@ -8,7 +8,7 @@
     "dependencies": {
         "Newtonsoft.Json": "6.0.6",
         "Microsoft.Framework.Configuration": "1.0.0-*",
-        "Microsoft.Framework.Configuration.File": "1.0.0-*"
+        "Microsoft.Framework.Configuration.FileExtensions": "1.0.0-*"
     },
     "frameworks": {
         "net45": { },
