+++ test/Microsoft.Framework.Configuration.FileExtensions.Test/FileConfigurationBuilderExtensionsTest.cs
