+++ test/Microsoft.Framework.Configuration.FileExtensions.Test/Microsoft.Framework.Configuration.FileExtensions.Test.xproj
